+++ Algoritmos/Simples/binary_search.c
