+++ Algoritmos/main.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "sort.h"
+#include "Complexo/sort.h"
 
 int comparator(void *item1, void *item2)
 {
