--- Estruturas/test/README.md
@@ -1,15 +1,0 @@
-# Description
-- Tool: Google Test
-- How to Download:
-
-```bash
-git clone https://github.com/google/googletest.git
-cd googletest
-mkdir build
-cd build
-cmake ..
-make
-sudo apt make install
-```
-
-- Observação: Só funciona ao escrever testes em arquivos .cpp
