+++ Listas/Nilton/Lista_Manipulação_arquivos/ler_espaço_branco.c
