+++ Listas/Ribas/COPA/COPA1-prof.c
