+++ Listas/Ribas/COPA/COPA1.c
