+++ Listas/Ribas/COPA/COPA2.c
