+++ Listas/Ribas/COPA/COPA3.c
