+++ Listas/Ribas/Lista 1/C/Ansiedade.c
