+++ Listas/Ribas/Lista 1/Cpp/HelloWorld2.cpp
