+++ Listas/Ribas/Lista 1/Cpp/Somar2.cpp
