+++ Listas/Ribas/Lista 1/Cpp/SomarN.cpp
