+++ Listas/Ribas/Lista 2/Aparecido.c
