+++ Listas/Ribas/Lista 2/AparecidoSimples.c
