+++ Listas/Ribas/Lista_4/inverse.c
